requirements.txt
pandas==2.1.4
python-dotenv==1.0.0
seaborn==0.13.0
seleniumbase==4.22.2
tqdm==4.66.1
